filehandling.py
This file contains the basics of file handling operations.
Opening a file :  opens a file.
Syntax : open( filename, mode)
There are three modes in file handling.
Writing :
This mode is used to write into a file if there is no file with given name then it will create a file. If the is an existing file it will be overwritten.
Reading:
This mode will read the content in files.
Appending:
This mode will add data into an existing file.

These are the operations we  use in the creating keylogger program.